requirements.txt
paho-mqtt>=2.0.0
kornia==0.7.1
matplotlib==3.8.2
numpy==1.26.4
opencv_python==4.9.0.80
opencv_python_headless==4.9.0.80
pytorch_lightning==2.1.3
scipy==1.12.0
sorcery==0.2.2
thop==0.1.1.post2209072238
torchmetrics==1.3.2
